Hide mentions header if there are no mentions
This is a terrible, terrible way to solve this problem and it hurts. The
parent should not have to keep a state to count the children of a list,
that's just really weird and requires a lot of boilerplate to have this
be the default way of working for lists with cached data. Actually, it's
kind of weird to count the children of the list in the way it works now,
as that breaks any type of component that fetches the data shown in the
list. So this should not be done in this way, and I should probably have
added more warning comments, since nobody will ever read this commit
message. 😞