Commit the package-lock file
Yes, I did that. I'm sorry. I know I'll regret this (or at least someone else
will regret it on my behalf) when we get a gigantic merge conflict on this file.
But at this point, I just wanted to say that I value being able to build the
project on a new install with the same dependencies as on everyone elses machine
a little bit more than the slight hassle of having a wacky merge conflict once
in a while. The previous commit had 9 files changed only because react-bootstrap
upgraded from 1.0.0 to 1.0.1. Yes, that's a terrible use of semantic versioning,
but it happened. (Also, something about binary vs API compatiblity in a TypeScript/
JavaScript world.) It's going to happen again and I'd rather have everyone
writing code against the same API than just slightly different ones. I hope that
this satisfies the person who just clicked blame to see who the heck thought it
was a good idea to commit this file. There are good reasons for this.