Fix issue with links to users in feed
This was a problem on the 'New comment ...' relationship message in the
feeds. The links to the user in these messages was broken because the DB
guid was used instead of the base64 application IDs, which happened
because in this particular case the ID was not converted on the backend.

Fixes #119